backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Auth & Security
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
passlib[argon2]==1.7.4

# Validation & Serialization
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Redis
redis[hiredis]==5.2.1

# HTTP Client (for Council runner)
httpx==0.28.1

# AI Providers (for Council identities)
anthropic==0.43.0
openai==1.58.1

# Content Processing
bleach==6.2.0
markdown==3.7

# Logging
structlog==24.4.0

# Utilities
python-multipart==0.0.20
python-dotenv==1.0.1
